@@ -1,113 +1,109 @@
-import 'package:flutter/material.dart';
+import 'dart:async';
 
-class TextInputWidget extends StatelessWidget {
-  final double width, height, fontSize, borderRadius;
-  final Color placeholderColor, inputTextColor, backgroundColor;
-  final EdgeInsets margin, marginWithError, padding;
-  final String placeholder;
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+class TextInputWidget extends StatefulWidget {
+  final double? width, height;
+  final EdgeInsets margin, marginWithError;
+  final EdgeInsets? padding;
   final VoidCallback? onTap;
+  final String? Function(String?)? onChanged;
   final TextEditingController? controller;
-  final String? errorMessage;
-  final bool obscureText;
+  final String? error, placeholder;
+  final bool obscureText, readOnly;
   final Widget? helpWidget;
 
   const TextInputWidget({
     super.key,
-    required this.placeholder,
-    this.width = 278,
+    this.placeholder,
+    this.width,
     this.height = 42,
-    this.borderRadius = 8,
-    this.fontSize = 17,
     this.marginWithError = const EdgeInsets.all(0),
     this.margin = const EdgeInsets.all(0),
-    this.padding = const EdgeInsets.only(left: 16, right: 16),
-    this.placeholderColor = const Color(0xFF8E8E93),
-    this.inputTextColor = Colors.black,
-    this.backgroundColor = Colors.white,
+    this.padding,
     this.onTap,
     this.controller,
-    this.errorMessage,
+    this.error,
     this.obscureText = false,
     this.helpWidget,
+    this.readOnly = false,
+    this.onChanged,
   });
 
   @override
+  State<TextInputWidget> createState() => _TextInputWidgetState();
+}
+
+class _TextInputWidgetState extends State<TextInputWidget> {
+  Timer? _debounce;
+
+  void _onTextChanged(String value) {
+    if (_debounce?.isActive ?? false) _debounce?.cancel();
+    _debounce = Timer(const Duration(milliseconds: 300), () {
+      widget.onChanged?.call(value);
+    });
+  }
+
+  @override
+  void dispose() {
+    _debounce?.cancel();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    const marginBackground = EdgeInsets.only(top: 3);
-    const errorColor = Colors.red;
-    const borderWidth = 1.0,
-        errorLeftPadding = 8.0,
-        errorTopPadding = 2.0,
-        errorFontSize = 12.0,
-        widthErrorMessageWithHelpWidget = 150.0;
+    const errorLeftPadding = 8.0;
+    final theme = Theme.of(context).inputDecorationTheme;
+    final effectiveBorder =
+        widget.error == null ? theme.border : theme.errorBorder;
+    final textStyleError = theme.errorStyle;
+    final marginContent =
+        widget.error == null ? widget.margin : widget.marginWithError;
+    final padding = theme.contentPadding;
     return Container(
-      margin: errorMessage == null ? margin : marginWithError,
-      child: Stack(
+      margin: marginContent,
+      child: Column(
         children: [
-          Column(
-            children: [
-              Container(
-                margin: marginBackground,
-                width: width,
-                height: height,
-                decoration: BoxDecoration(
-                  color: backgroundColor,
-                  borderRadius: BorderRadius.circular(borderRadius),
-                  border: errorMessage != null
-                      ? Border.all(color: errorColor, width: borderWidth)
-                      : null,
-                ),
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Container(
-                    alignment: Alignment.centerLeft,
-                    padding: EdgeInsets.only(
-                        left: errorLeftPadding, top: errorTopPadding),
-                    child: errorMessage != null
-                        ? SizedBox(
-                            width: helpWidget == null
-                                ? width - errorLeftPadding * 2
-                                : widthErrorMessageWithHelpWidget,
-                            child: Text(
-                              errorMessage!,
-                              style: TextStyle(
-                                fontSize: errorFontSize,
-                                color: errorColor,
-                              ),
-                            ))
-                        : null,
-                  ),
-                  Container(child: errorMessage != null ? helpWidget : null),
-                ],
-              ),
-              Container(child: errorMessage == null ? helpWidget : null),
-            ],
-          ),
-          Container(
-            padding: padding,
+          SizedBox(
+            height: widget.height,
+            width: widget.width,
             child: TextFormField(
-              controller: controller,
-              onTap: onTap,
-              obscureText: obscureText,
-              style: TextStyle(
-                color: inputTextColor,
-                fontSize: fontSize,
-                fontWeight: FontWeight.normal,
-              ),
+              inputFormatters: [
+                FilteringTextInputFormatter.allow(RegExp(r'[a-zA-Z0-9@._-]'))
+              ],
+              readOnly: widget.readOnly,
+              controller: widget.controller,
+              obscureText: widget.obscureText,
+              onChanged: _onTextChanged,
               decoration: InputDecoration(
-                hintText: placeholder,
-                hintStyle: TextStyle(
-                  color: placeholderColor,
-                  fontSize: fontSize,
-                  fontWeight: FontWeight.normal,
-                ),
-                border: InputBorder.none,
-                enabledBorder: InputBorder.none,
-                focusedBorder: InputBorder.none,
+                contentPadding: padding ?? EdgeInsets.symmetric(horizontal: 16),
+                hintText: widget.placeholder,
+                border: effectiveBorder,
+                enabledBorder: effectiveBorder,
+                focusedBorder: effectiveBorder,
               ),
             ),
+          ),
+          Row(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisAlignment: MainAxisAlignment.end,
+            children: [
+              if (widget.error != null)
+                Flexible(
+                  child: Container(
+                    alignment: Alignment.centerLeft,
+                    padding: EdgeInsets.symmetric(horizontal: errorLeftPadding),
+                    child: Text(
+                      widget.error!,
+                      style: textStyleError,
+                      softWrap: true,
+                      overflow: TextOverflow.visible,
+                    ),
+                  ),
+                ),
+              if (widget.helpWidget != null) widget.helpWidget!,
+            ],
           ),
         ],
       ),
