@@ -2,9 +2,11 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:receptico/core/router/router.dart';
+import 'package:receptico/generated/l10n.dart';
 
-import '../block/auth_block.dart';
-import '../model/user.dart';
+import '../bloc/auth_bloc.dart';
+import '../common/enum.dart';
+import '../common/mixin.dart';
 import '../widget/widget.dart';
 
 @RoutePage()
@@ -15,156 +17,115 @@ class RegisterPage extends StatefulWidget {
   State<RegisterPage> createState() => _RegisterPageState();
 }
 
-class _RegisterPageState extends State<RegisterPage> {
-  late bool _isError;
-  AuthRegisterFailure _authState = AuthRegisterFailure();
-  final _usernameController = TextEditingController();
-  final _phoneController = TextEditingController();
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
+class _RegisterPageState extends State<RegisterPage>
+    with ValidateMixin, ShowTimerDialogueMixin {
+  final Map<EInput, TextEditingController> _controllers = {
+    EInput.username: TextEditingController(),
+    EInput.emailOrPhone: TextEditingController(),
+    EInput.password: TextEditingController(),
+  };
+
+  void _clearForm() =>
+      _controllers.forEach((key, controller) => controller.clear());
 
   @override
-  void initState() {
-    super.initState();
-    _isError = false;
-  }
-
-  void _clearFailMessage() {
-    if (_isError) {
-      _isError = false;
-      setState(() => _authState = AuthRegisterFailure());
-    }
+  void dispose() {
+    _controllers.forEach((key, controller) => controller.dispose());
+    super.dispose();
   }
 
   void _submit() {
+    usernameValidate(_controllers[EInput.username]?.text);
+    emailValidate(_controllers[EInput.emailOrPhone]?.text);
+    passwordValidate(_controllers[EInput.password]?.text);
+
     context.read<AuthBloc>().add(
-          AuthRegister(
-            user: RegisterUser(
-              email: _emailController.text,
-              username: _usernameController.text,
-              phone: _phoneController.text,
-              password: _passwordController.text,
-            ),
+          AuthRegisterEvent(
+            username: _controllers[EInput.username]?.text ?? '',
+            email: _controllers[EInput.emailOrPhone]?.text ?? '',
+            password: _controllers[EInput.password]?.text ?? '',
           ),
         );
   }
 
   @override
   Widget build(BuildContext context) {
-    const marginTextInput = EdgeInsets.only(bottom: 10.0),
-        marginTextInputWithError = EdgeInsets.only(bottom: 7.0);
-    var router = AutoRouter.of(context);
+    final router = AutoRouter.of(context);
+    final localization = S.of(context);
     return BlocListener<AuthBloc, AuthState>(
       listener: (context, state) {
-        if (state is AuthRegisterSuccess) {
-          router.goTo(StartRoute());
-        }
-        if (state is AuthRegisterFailure) {
-          _isError = true;
-          setState(() => _authState = state);
-          return;
+        switch (state.runtimeType) {
+          case const (AuthClearFailState):
+            _clearForm();
+            break;
+
+          case const (AuthRegisterSuccessState):
+            router.navigate(const SendEmailRoute());
+            break;
+
+          case const (AuthShowWaitMessageState):
+            showTimedDialog(context);
+            break;
         }
       },
-      child: Scaffold(
+      child: ScaffoldWithGradientWidget(
         body: Stack(
           children: [
-            RegisterScreenWidget(),
+            const ScreenBackgroundWidget(IsMirrored: true),
             Container(
               alignment: Alignment.center,
               child: SizedBox(
                 width: 278,
-                height: 520,
+                height: 430,
                 child: SingleChildScrollView(
                   child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.stretch,
                     children: [
-                      const TitleWidget(
-                        title: 'Реєстрація',
-                        subtitle: 'Створіть новий акаунт',
+                      TitleWidget(
+                        title: localization.registerTitle,
+                        subtitle: localization.registerSubtitle,
                       ),
                       const SizedBox(height: 13),
-                      TextInputWidget(
-                        placeholder: 'Ім’я',
-                        controller: _usernameController,
-                        onTap: _clearFailMessage,
-                        errorMessage: _authState.usernameError,
-                        margin: marginTextInput,
-                        marginWithError: marginTextInputWithError,
+                      SelectorTextInputWidget(
+                        placeholder: localization.namePlaceholder,
+                        controller: _controllers[EInput.username],
+                        selector: (state) => state.errors?[EBlocError.username],
+                        onChanged: usernameValidate,
                       ),
-                      TextInputWidget(
-                        placeholder: 'Email',
-                        controller: _emailController,
-                        onTap: _clearFailMessage,
-                        errorMessage: _authState.emailError,
-                        margin: marginTextInput,
-                        marginWithError: marginTextInputWithError,
+                      SelectorTextInputWidget(
+                        placeholder: localization.emailPlaceholder,
+                        controller: _controllers[EInput.emailOrPhone],
+                        selector: (state) => state.errors?[EBlocError.email],
+                        onChanged: emailValidate,
                       ),
-                      TextInputWidget(
-                        placeholder: 'Номер телефону',
-                        controller: _phoneController,
-                        onTap: _clearFailMessage,
-                        errorMessage: _authState.phoneError,
-                        margin: marginTextInput,
-                        marginWithError: marginTextInputWithError,
+                      SelectorPasswordInputWidget(
+                        placeholder: localization.passwordPlaceholder,
+                        controller: _controllers[EInput.password],
+                        selector: (state) => state.errors?[EBlocError.password],
+                        onChanged: passwordValidate,
                       ),
-                      PasswordInputWidget(
-                        controller: _passwordController,
-                        placeholderText: 'Пароль',
-                        onTap: _clearFailMessage,
-                        errorMessage: _authState.passwordError,
-                        margin: const EdgeInsets.only(bottom: 25.0),
-                        marginWithError: const EdgeInsets.only(bottom: 22.0),
-                      ),
-                      LinkButtonWidget(
-                        text: 'Створити',
-                        onPressed: _submit,
+                      SizedBox(
+                        height: 50,
+                        child: TextButton(
+                          onPressed: _submit,
+                          child: Text(localization.registerButton),
+                        ),
                       ),
                       const SizedBox(height: 16),
                       FooterWidget(
-                        text: 'Вже маєте акаунт?',
-                        linkText: 'Увійти',
-                        onTab: () {
-                          router.goTo(LoginRoute());
-                          _clearFailMessage();
-                        },
+                        text: localization.accountQuestion,
+                        linkText: localization.loginLink,
+                        onTab: () => router.navigate(const LoginRoute()),
                       ),
                     ],
                   ),
                 ),
               ),
             ),
+            LoadingWidget(),
           ],
         ),
       ),
-    );
-  }
-}
-
-class RegisterScreenWidget extends StatelessWidget {
-  const RegisterScreenWidget({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      fit: StackFit.expand,
-      children: [
-        const Positioned(
-          top: -95,
-          right: -76,
-          child: CircleWidget(width: 254, height: 254),
-        ),
-        const Positioned(
-          bottom: -34,
-          right: -37,
-          child: CircleWidget(width: 123, height: 123),
-        ),
-        const Positioned(
-          bottom: -71,
-          left: -33,
-          child: CircleWidget(width: 196, height: 196),
-        ),
-      ],
     );
   }
 }
