@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:receptico/core/UI/theme.dart';
 
 class FooterWidget extends StatelessWidget {
+  final double spacing;
   final String text, linkText;
   final VoidCallback? onTab;
 
@@ -8,23 +10,22 @@ class FooterWidget extends StatelessWidget {
     super.key,
     required this.text,
     required this.linkText,
+    this.spacing = 8.0,
     this.onTab,
   });
 
   @override
   Widget build(BuildContext context) {
-    const spaceWidth = 8.0;
+    final textStyleLink = context.font.subheadBold,
+        textStyleText = context.font.subhead;
     return Row(
       mainAxisAlignment: MainAxisAlignment.center,
+      spacing: spacing,
       children: [
-        Text(text),
-        const SizedBox(width: spaceWidth),
+        Text(text, style: textStyleText),
         InkWell(
           onTap: onTab,
-          child: Text(
-            linkText,
-            style: const TextStyle(fontWeight: FontWeight.bold),
-          ),
+          child: Text(linkText, style: textStyleLink),
         ),
       ],
     );
